+++ pkg/agents/team1/agent1/agent.go
@@ -1,4 +1,4 @@
-package agent
+package agent1
 
 import (
 	"log"
