+++ pkg/agents/team1/agent2/agent.go
@@ -1,4 +1,4 @@
-package agent
+package agent2
 
 import (
 	"log"
